feat: check map surrounded by wall, norm #5
