+++ check_argv/degree_utils.c
@@ -2,7 +2,7 @@
 
 double	deg2rad(double degree)
 {
-	return ((degree) * PI / 180.0); 
+	return ((degree) * PI / 180.0);
 }
 
 double	adjust_degree(double base_degree, double offset_degree)
