+++ check_argv/make_ceiling_floor_image.c
@@ -2,33 +2,36 @@
 #include "../cub3d.h"
 #include <stdio.h>
 
-#define START	0
-#define END		1
+typedef struct outside_wall_info
+{
+	int	x;
+	int	range_y[2];
+}	t_outside_wall_info;
 
-static void	_pixel_put_to_img(t_bg_data *bg_data, int x, int range_y[2], \
-										int color, double base_transperency)
+static void	_pixel_put_to_img(t_bg_data *bg_data, int color, \
+					t_outside_wall_info outside_wall, double base_transperency)
 {
 	char	*dst;
-	int		y;
 	int		step;
 	int		sign;
-	double		gradiant;
+	int		y;
+	double	gradiant;
 
-	step = range_y[END] - range_y[START];
+	step = outside_wall.range_y[END] - outside_wall.range_y[START];
 	sign = 1;
 	if (step < 0)
 	{
 		sign = -1;
 		step *= -1;
 	}
-	y = range_y[START];
 	gradiant = base_transperency;
-	while (y != range_y[END])
+	y = outside_wall.range_y[START];
+	while (y != outside_wall.range_y[END])
 	{
 		gradiant += (1.0 - base_transperency) / step;
-		dst = bg_data->addr + (y * bg_data->line_length + x \
-									* (bg_data->bits_per_pixel / 8));
-		*(unsigned int*)dst = create_trgb(0, get_r(color) * gradiant, \
+		dst = bg_data->addr + (y * bg_data->line_length + \
+								outside_wall.x * (bg_data->bits_per_pixel / 8));
+		*(unsigned int *)dst = create_trgb(0, get_r(color) * gradiant, \
 							get_g(color) * gradiant, get_b(color) * gradiant);
 		y += sign;
 	}
@@ -46,19 +49,22 @@ static void	_set_range_y(int y[2], int y_start, int y_end)
 	y[END] = y_end;
 }
 
-void	draw_ceiling_floor(t_game *game, int x, int y_top, int y_bottom, double fog_value)
+void	draw_ceiling_floor(t_game *game, int x, int wall_y[2], double fog_value)
 {
-	const int	screen_y = game->info.screen_y;
-	int			y[2];
+	const int			screen_y = game->info.screen_y;
+	t_outside_wall_info	outside_wall;
 
-	if (y_top > 0)
+	outside_wall.x = x;
+	if (wall_y[START] > 0)
 	{
-		_set_range_y(y, y_top, 0);
-		_pixel_put_to_img(&game->bg_data, x, y, game->ceiling_color, fog_value);
+		_set_range_y(outside_wall.range_y, wall_y[START], 0);
+		_pixel_put_to_img(&game->bg_data, game->ceiling_color, \
+										outside_wall, fog_value);
 	}
-	if (y_bottom < screen_y)
+	if (wall_y[END] < screen_y)
 	{
-		_set_range_y(y, y_bottom, screen_y);
-		_pixel_put_to_img(&game->bg_data, x, y, game->floor_color, fog_value);
+		_set_range_y(outside_wall.range_y, wall_y[END], screen_y);
+		_pixel_put_to_img(&game->bg_data, game->floor_color, \
+										outside_wall, fog_value);
 	}
 }
