+++ check_argv/rgb_utils.c
@@ -1,4 +1,4 @@
-int create_trgb(unsigned char t, unsigned char r, \
+int	create_trgb(unsigned char t, unsigned char r, \
 								unsigned char g, unsigned char b)
 {
 	return (*(int *)(unsigned char [4]){b, g, r, t});
