+++ check_argv/sample_map.cub
@@ -26,4 +26,4 @@ C 225,10,2
 10000000000000001101010010001
 1100000111010101111101111N00111
 11110111 1110101 101111010001
-11111111 1111111 111111111111
+11111111 1111111 1111111111111
